# public/swagger/swagger-doc.json
{
  "openapi": "2.0.0",
  "info": {
    "description": "数据管理平台",
    "version": "1.0.0",
    "title": "数据管理平台-接口文档",
    "contact": {
      "name": "async",
      "email": ""
    },
    "termsOfService": "http://www.yczcjk.com/",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "paths": {
    "/login/in": {
      "post": {
        "tags": [
          "登录模块"
        ],
        "summary": "Add a new pet to the store",
        "operationId": "loginInGet1",
        "parameters": [
          {
            "in": "body",
            "name": "info",
            "required": true,
            "type": "object",
            "description": "其他相关信息",
            "schema": {
              "$ref": "#/definitions/Order"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "$ref": "#/definitions/Order"
            }
          },
          "201": {
            "description": "error"
          },
          "400": {
            "description": "no-111111"
          }
        }
      }
    }
  },
  "definitions": {
    "Order": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64",
          "example": 85,
          "description": "有效标识（0：失效，1：有效）"
        },
        "petId": {
          "type": "integer",
          "format": "int64"
        },
        "quantity": {
          "type": "integer",
          "format": "int32"
        },
        "shipDate": {
          "type": "string",
          "format": "date-time"
        },
        "status": {
          "type": "string",
          "description": "Order Status",
          "enum": [
            "placed",
            "approved",
            "delivered"
          ]
        },
        "complete": {
          "type": "boolean",
          "default": false
        }
      },
      "xml": {
        "name": "Order"
      }
    }
  },
  "components": {},
  "tags": []
}